+++ Sources/KeyboardKit/Layout/StaticKeyboardLayoutProvider.swift
@@ -10,10 +10,7 @@ import Foundation
 
 /**
  This keyboard layout provider returns the layout that it is
- initialized with.
- 
- It can be used to return a static layout, without taking in
- factors like device and orientation into consideration.
+ initialized with, without any additional logic.
  */
 public class StaticKeyboardLayoutProvider: KeyboardLayoutProvider {